requirements.txt
numpy>=1.24
pandas>=2.1
scipy>=1.11
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.13
statsmodels>=0.14
dtaidistance>=2.3
tslearn>=0.6
pyyaml>=6.0
tqdm>=4.65
networkx>=3.1
pillow>=10.0
torch>=2.1
torchvision>=0.16
